+++ Userland/Libraries/LibWeb/CSS/ComputedValues.h
@@ -8,13 +8,14 @@
 
 #include <AK/Optional.h>
 #include <LibWeb/CSS/BackdropFilter.h>
+#include <LibWeb/CSS/CalculatedOr.h>
 #include <LibWeb/CSS/Clip.h>
 #include <LibWeb/CSS/Display.h>
 #include <LibWeb/CSS/GridTrackPlacement.h>
 #include <LibWeb/CSS/GridTrackSize.h>
 #include <LibWeb/CSS/LengthBox.h>
+#include <LibWeb/CSS/PercentageOr.h>
 #include <LibWeb/CSS/Size.h>
-#include <LibWeb/CSS/StyleValue.h>
 #include <LibWeb/CSS/StyleValues/AbstractImageStyleValue.h>
 #include <LibWeb/CSS/StyleValues/ShadowStyleValue.h>
 #include <LibWeb/CSS/TransformFunctions.h>
