+++ Userland/Libraries/LibWeb/CSS/LengthBox.cpp
@@ -6,7 +6,6 @@
  */
 
 #include "LengthBox.h"
-#include <LibWeb/CSS/StyleValue.h>
 
 namespace Web::CSS {
