+++ Userland/Libraries/LibWeb/CSS/Position.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <LibWeb/CSS/Percentage.h>
+#include <LibWeb/CSS/PercentageOr.h>
 #include <LibWeb/PixelUnits.h>
 
 namespace Web::CSS {
