+++ Userland/Libraries/LibWeb/CSS/Resolution.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "Resolution.h"
-#include <LibWeb/CSS/StyleValue.h>
 
 namespace Web::CSS {
