+++ Userland/Libraries/LibWeb/CSS/Size.cpp
@@ -4,9 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <LibWeb/CSS/Length.h>
 #include <LibWeb/CSS/Size.h>
-#include <LibWeb/CSS/StyleValue.h>
 
 namespace Web::CSS {
