+++ Userland/Libraries/LibWeb/CSS/Size.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <LibWeb/CSS/Length.h>
-#include <LibWeb/CSS/Percentage.h>
+#include <LibWeb/CSS/PercentageOr.h>
 
 namespace Web::CSS {
