+++ Userland/Libraries/LibWeb/CSS/StyleValues/LinearGradientStyleValue.h
@@ -10,6 +10,8 @@
 #pragma once
 
 #include <AK/Vector.h>
+#include <LibWeb/CSS/Angle.h>
+#include <LibWeb/CSS/Percentage.h>
 #include <LibWeb/CSS/StyleValues/AbstractImageStyleValue.h>
 #include <LibWeb/Painting/GradientPainting.h>
