+++ Userland/Libraries/LibWeb/HTML/HTMLBodyElement.cpp
@@ -5,7 +5,6 @@
  */
 
 #include <LibWeb/CSS/StyleProperties.h>
-#include <LibWeb/CSS/StyleValue.h>
 #include <LibWeb/CSS/StyleValues/ColorStyleValue.h>
 #include <LibWeb/CSS/StyleValues/ImageStyleValue.h>
 #include <LibWeb/DOM/Document.h>
